requirements.txt
selenium==4.21.0
pytest==8.2.0
pytest-xdist==3.5.0
allure-pytest==2.13.5
